# Avoid rendering during timeline replay